src/Y2MP3/pyproject.toml
[build-system]
requires = ["briefcase"]
build-backend = "briefcase"

[project]
name = "Y2MP3"
version = "1.0.0"
description = "تحميل وتحويل الفيديوهات"
author = "ALI ALBEDI"
dependencies = ["yt-dlp", "requests", "toga-android~=0.5.0"]

[tool.briefcase.app.Y2MP3]
formal_name = "Y2MP3-Pro"
bundle = "com.drali.y2mp3pro"
description = "تطبيق أندرويد لتحميل وتحويل الفيديوهات"
sources = ["src/Y2MP3"]

[tool.briefcase.app.Y2MP3.android]
requires = ["yt-dlp", "requests", "toga-android~=0.5.0"]